Move verify.go to lokomotive package.
- install package consists of two files:
  install.go: which deals with configuring terraform.
  verify.go: which deals with verification of Lokomotive cluster.

  install package by its name and contents of the files serve no
  purpose, hence its files should be moved around to correct packages.

- This commit deals with moving the verify.go file to lokomotive
  package.
